LPS-166857 Moves converters to layout-api
