+++ modules/apps/headless/headless-delivery/headless-delivery-impl/src/main/java/com/liferay/headless/delivery/internal/dto/v1_0/mapper/CollectionLayoutStructureItemMapper.java
@@ -25,10 +25,10 @@ import com.liferay.headless.delivery.dto.v1_0.PageCollectionDefinition;
 import com.liferay.headless.delivery.dto.v1_0.PageElement;
 import com.liferay.info.list.provider.item.selector.criterion.InfoListProviderItemSelectorReturnType;
 import com.liferay.item.selector.criteria.InfoListItemSelectorReturnType;
-import com.liferay.layout.page.template.util.AlignConverter;
-import com.liferay.layout.page.template.util.FlexWrapConverter;
-import com.liferay.layout.page.template.util.JustifyConverter;
 import com.liferay.layout.responsive.ViewportSize;
+import com.liferay.layout.util.converter.AlignConverter;
+import com.liferay.layout.util.converter.FlexWrapConverter;
+import com.liferay.layout.util.converter.JustifyConverter;
 import com.liferay.layout.util.structure.CollectionStyledLayoutStructureItem;
 import com.liferay.layout.util.structure.LayoutStructureItem;
 import com.liferay.layout.util.structure.collection.EmptyCollectionOptions;
