+++ modules/apps/headless/headless-delivery/headless-delivery-impl/src/main/java/com/liferay/headless/delivery/internal/dto/v1_0/mapper/FormLayoutStructureItemMapper.java
@@ -27,10 +27,10 @@ import com.liferay.headless.delivery.dto.v1_0.URLFormSubmissionResult;
 import com.liferay.headless.delivery.internal.dto.v1_0.mapper.util.FragmentMappedValueUtil;
 import com.liferay.headless.delivery.internal.dto.v1_0.mapper.util.LocalizedValueUtil;
 import com.liferay.headless.delivery.internal.dto.v1_0.mapper.util.StyledLayoutStructureItemUtil;
-import com.liferay.layout.page.template.util.AlignConverter;
-import com.liferay.layout.page.template.util.ContentDisplayConverter;
-import com.liferay.layout.page.template.util.FlexWrapConverter;
-import com.liferay.layout.page.template.util.JustifyConverter;
+import com.liferay.layout.util.converter.AlignConverter;
+import com.liferay.layout.util.converter.ContentDisplayConverter;
+import com.liferay.layout.util.converter.FlexWrapConverter;
+import com.liferay.layout.util.converter.JustifyConverter;
 import com.liferay.layout.util.structure.FormStyledLayoutStructureItem;
 import com.liferay.layout.util.structure.LayoutStructureItem;
 import com.liferay.portal.kernel.json.JSONObject;
