+++ modules/apps/layout/layout-api/src/main/java/com/liferay/layout/util/converter/PaddingConverter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.page.template.util;
+package com.liferay.layout.util.converter;
 
 import com.liferay.portal.kernel.util.HashMapBuilder;
 
@@ -21,17 +21,25 @@ import java.util.Objects;
 import java.util.Set;
 
 /**
- * @author Eudaldo Alonso
+ * @author Rubén Pulido
  */
-public class ContentDisplayConverter {
+public class PaddingConverter {
 
 	public static final Map<String, String> externalToInternalValuesMap =
 		HashMapBuilder.put(
-			"Block", "block"
+			"0", "0"
 		).put(
-			"FlexColumn", "flex-column"
+			"1", "3"
 		).put(
-			"FlexRow", "flex-row"
+			"2", "4"
+		).put(
+			"4", "5"
+		).put(
+			"6", "6"
+		).put(
+			"8", "7"
+		).put(
+			"10", "8"
 		).build();
 
 	public static String convertToExternalValue(String value) {
