+++ modules/apps/layout/layout-api/src/main/java/com/liferay/layout/util/converter/ShadowConverter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.page.template.util;
+package com.liferay.layout.util.converter;
 
 import com.liferay.portal.kernel.util.HashMapBuilder;
