+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/importer/structure/util/ContainerLayoutStructureItemImporter.java
@@ -17,16 +17,16 @@ package com.liferay.layout.internal.importer.structure.util;
 import com.liferay.headless.delivery.dto.v1_0.FragmentLink;
 import com.liferay.headless.delivery.dto.v1_0.PageElement;
 import com.liferay.layout.internal.importer.LayoutStructureItemImporterContext;
-import com.liferay.layout.page.template.util.AlignConverter;
-import com.liferay.layout.page.template.util.BorderRadiusConverter;
-import com.liferay.layout.page.template.util.ContentDisplayConverter;
-import com.liferay.layout.page.template.util.ContentVisibilityConverter;
-import com.liferay.layout.page.template.util.FlexWrapConverter;
-import com.liferay.layout.page.template.util.HtmlTagConverter;
-import com.liferay.layout.page.template.util.JustifyConverter;
-import com.liferay.layout.page.template.util.MarginConverter;
-import com.liferay.layout.page.template.util.PaddingConverter;
-import com.liferay.layout.page.template.util.ShadowConverter;
+import com.liferay.layout.util.converter.AlignConverter;
+import com.liferay.layout.util.converter.BorderRadiusConverter;
+import com.liferay.layout.util.converter.ContentDisplayConverter;
+import com.liferay.layout.util.converter.ContentVisibilityConverter;
+import com.liferay.layout.util.converter.FlexWrapConverter;
+import com.liferay.layout.util.converter.HtmlTagConverter;
+import com.liferay.layout.util.converter.JustifyConverter;
+import com.liferay.layout.util.converter.MarginConverter;
+import com.liferay.layout.util.converter.PaddingConverter;
+import com.liferay.layout.util.converter.ShadowConverter;
 import com.liferay.layout.util.structure.ContainerStyledLayoutStructureItem;
 import com.liferay.layout.util.structure.LayoutStructure;
 import com.liferay.layout.util.structure.LayoutStructureItem;
