+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/importer/structure/util/FormLayoutStructureItemImporter.java
@@ -17,10 +17,10 @@ package com.liferay.layout.internal.importer.structure.util;
 import com.liferay.headless.delivery.dto.v1_0.ContextReference;
 import com.liferay.headless.delivery.dto.v1_0.PageElement;
 import com.liferay.layout.internal.importer.LayoutStructureItemImporterContext;
-import com.liferay.layout.page.template.util.AlignConverter;
-import com.liferay.layout.page.template.util.ContentDisplayConverter;
-import com.liferay.layout.page.template.util.FlexWrapConverter;
-import com.liferay.layout.page.template.util.JustifyConverter;
+import com.liferay.layout.util.converter.AlignConverter;
+import com.liferay.layout.util.converter.ContentDisplayConverter;
+import com.liferay.layout.util.converter.FlexWrapConverter;
+import com.liferay.layout.util.converter.JustifyConverter;
 import com.liferay.layout.util.structure.FormStyledLayoutStructureItem;
 import com.liferay.layout.util.structure.LayoutStructure;
 import com.liferay.layout.util.structure.LayoutStructureItem;
